requirements.txt
streamlit
pycognito
websocket-client